feat: support for qualified module names
BREAKING
When a module is registered with `asap_declare_module()`, its MODULE_NAME
can be composed of multiple segments separated by '.' or '_'.
In such case, these segments will be used as path segments for the `api_export`
generated file, and as identifier segments in the corresponding `_API` compiler
defines.

For example, `Super.Hero.Module` will produce a file that can be included as
"super/hero/module/api_export.h" and will provide the export macro as
`SUPER_HERO_MODULE_API`.